Renamed LinkMediaTracking to LinkReferenceFactory
