+++ src/Skybrud.LinkPicker/Composers/LinkPickerComposer.cs
@@ -1,4 +1,4 @@
-﻿using Skybrud.LinkPicker.Tracking;
+﻿using Skybrud.LinkPicker.Factories;
 using Umbraco.Core;
 using Umbraco.Core.Composing;
 
@@ -7,7 +7,7 @@ namespace Skybrud.LinkPicker.Composers {
     public class LinkPickerComposer : IUserComposer {
         
         public void Compose(Composition composition) {
-            composition.DataValueReferenceFactories().Append<LinkMediaTracking>();
+            composition.DataValueReferenceFactories().Append<LinkReferenceFactory>();
         }
 
     }
